+++ CHANGELOG.md
@@ -30,6 +30,7 @@ All notable changes to this project will be documented in this file.
 - #1643 Separate AS1/2 and AS3 editor hilighting
 - AS3 direct editation - slot/const default values
 - #1328 AS1/2/3 direct editation - empty commands (just semicolon)
+- #1310 AS1/2/3 direct editation - modulo operator precedence
 
 ### Removed
 - #1631 ActiveX Flash component download in windows installer